requirements.txt
asgiref
Django
django-allauth==0.54
djangorestframework
gunicorn
oauthlib
psycopg2-binary
requests-oauthlib==1.3.1
urllib3==1.26.11
django-cors-headers==3.13.0
dj-rest-auth==2.2.5
djangorestframework-simplejwt==5.2.1
python-dotenv
simple-salesforce